intel-ai-hackathon/requirements.txt
streamlit>=1.32.0
openai>=1.30.0
faiss-cpu>=1.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.18.0
spacy>=3.7.0
